pyproject.toml
[project]
name = "sales-analyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.3",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.6",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "weaviate-client>=4.19.2",
]
